build VRM10Viewer and SimpleViewer
