+++ Assets/VRM/Editor/VrmTopMenu.cs
@@ -53,13 +53,13 @@ namespace VRM
         private static void ShowVersionDialog() => VRMVersionMenu.ShowVersionDialog();
 
         [MenuItem(DevelopmentMenuPrefix + "/Build dummy for CI", false, 93)]
-        private static void BuildDummyForCi() => BuildClass.Build();
+        private static void BuildDummyForCi() => DevOnly.BuildClass.Build();
 
-        [MenuItem(DevelopmentMenuPrefix + "/Build 0x WebGL for CI", false, 94)]
-        private static void BuildWebGLForCi0x() => BuildClass.BuildWebGL_SimpleViewer();
+        [MenuItem(DevelopmentMenuPrefix + "/Switch to built-in RenderPipeline", false, 94)]
+        private static void SwitchToBuiltinPipeline() => DevOnly.BuildClass.SwitchBuiltinPipeline();
 
-        [MenuItem(DevelopmentMenuPrefix + "/Build 10 WebGL for CI", false, 95)]
-        private static void BuildWebGLForCi10() => BuildClass.BuildWebGL_VRM10Viewer();
+        [MenuItem(DevelopmentMenuPrefix + "/Build WebGL SimpleViewer for CI", false, 95)]
+        private static void BuildWebGLForCi0x() => DevOnly.BuildClass.BuildWebGL_SimpleViewer();
 
         [MenuItem(DevelopmentMenuPrefix + "/Create UnityPackage", false, 99)]
         private static void CreateUnityPackage() => VRMExportUnityPackage.CreateUnityPackageWithoutBuild();
