+++ Assets/VRM10/Editor/BuildClass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3783cb4bfaca889418092d0fdde301ad
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
